server: extract datagram dispatch from UdpServer
Everything UdpServer did with a received wire packet moves into a
carrier-independent `datagram::data_received`, except the two steps
only a socket can do. It parses and validates the header, then routes
to a connection by peer address and then by session id. It delivers,
and commits a moved peer address only after a successful decrypt.

The two remaining steps arrive as closures: minting the per-connection
send callback, and replying to a peer that has no connection. Those are
the only ones that need a socket, so that is where the seam goes.

No behaviour change.
