+++ lightway-server/src/io/outside.rs
@@ -1,3 +1,4 @@
+pub(crate) mod datagram;
 pub(crate) mod tcp;
 pub(crate) mod udp;
